prom_discovery/requirements.txt
requests
progressbar2
pysnmp
